feat: Add support for NetBox v4.x in Annet (related to annetutil/annet#247)
