+++ annet/adapters/netbox/provider.py
@@ -1,6 +1,6 @@
 from typing import Dict, Any, Optional
 
-from dataclass_rest.exceptions import ClientError
+from dataclass_rest.exceptions import ClientError, ClientLibraryError
 
 from annet.storage import StorageProvider, Storage
 from annet.connectors import AdapterWithName, AdapterWithConfig, T
@@ -9,21 +9,34 @@ from .common.storage_opts import NetboxStorageOpts
 from .common.query import NetboxQuery
 from .v24.storage import NetboxStorageV24
 from .v37.storage import NetboxStorageV37
+from .v41.storage import NetboxStorageV41
+from .v42.storage import NetboxStorageV42
 
 
 def storage_factory(opts: NetboxStorageOpts) -> Storage:
     client = NetboxStatusClient(opts.url, opts.token, opts.insecure)
+    version_class_map = {
+        "3.": NetboxStorageV37,
+        "4.0": NetboxStorageV41,
+        "4.1": NetboxStorageV41,
+        "4.2": NetboxStorageV42,
+    }
+
+    status = None
+
     try:
         status = client.status()
+        for version_prefix, storage_class in version_class_map.items():
+            if status.netbox_version.startswith(version_prefix):
+                return storage_class(opts)
+
     except ClientError as e:
         if e.status_code == 404:
-            # old version do not support status reqeust
             return NetboxStorageV24(opts)
-        raise
-    if status.netbox_version.startswith("3."):
-        return NetboxStorageV37(opts)
-    else:
-        raise ValueError(f"Unsupported version: {status.netbox_version}")
+        else:
+            raise ValueError(f"Unsupported version: {status.netbox_version}")
+    except ClientLibraryError:
+        raise ValueError(f"Connection error: Unable to reach Netbox at URL: {opts.url}")
 
 
 class NetboxProvider(StorageProvider, AdapterWithName, AdapterWithConfig):
